Move DealWithFees to impls
